Remove Python 3.9 specific code
I had issues with tests passing on my local machine but not in the GitHub Actions. I have since discovered the issue (I think) which has to do with using some incompatible 3.9 features involving typing for lists. This has since been resolved and my tests are also not the issue.