Created script for translation between braille and english, as requested: - Letters a-z (and capitilization) - Numbers 0-9
Note, according to the specs numbers and letters must appear in different blocks separated by spaces.

Also note, I left a bunch of commented out code. Although this may be unprofessional if this was being shipped immediately, I thought it'd be good for
helping the team understand how I think and how I work. Frankly, I think theres a lot of extensibility here. Cases to consider which I did not implement as it was not
in the constraints and I don't want to fail the unit tests outlined. Examples like allowing numbers and letters in the same block.

I have ideas for these, as commented above, but due to concerns for potentially altering it beyond requirements and failing unit tests I did not uncomment this code/implement.