Challenge Completed 🚀
What:

In the challenge README it is mentioned to "include the entire English alphabet, the ability to capitalize letters, add spaces, and the numbers 0 through 9 as well."However I decided to include the ability to use decimal nubers too. I did not include special characters because it was not possible to differentiate between the letter "o" and the character ">" since they have the same braille sequence.

As for what my program considers a valid input a include the following rules:
For Braille:
- Length  should be multiple of 6
- Should only contain a series of 0 and .
- Should exist in the considered Braille alphabet
For english:
- Letters from a-z A-Z
- Numbers from 0-9
* Special characters are not considered a valid input