finished program
I think because of the python version used its only works if translator.test.py is renamed to translator_test.py. Also, python is used to run instead of python3. I don't think this will be a problem when ran on the companys' system where they are using python 3.8 for some reason