Implement translation from english to braille
The EnglishToBrailleTranslator function grabs a lookup map from a separate function due to the verbosity of the map. It then follows  the rules layed out in the challenge to parse an English string into a braille string. Implemented translation as a function in order to keep tests executable just from translator.go file without needing dependencies to other files (perhaps in the case of a OO approach).