Removed < and > due to Braille code conflict with "o"
Removed the characters < and >, as < shares the same Braille code as the letter "o" in the provided diagram. This was causing issues when converting Braille to English, particularly in correctly printing the letter "o". Since most special characters are not required for the challenge, I decided to remove them rather than overcomplicate the code. However, I kept the special characters that don't interfere with the translation process.